feat: 게시글 상세 조회, 유효성 검증 추가

+++ src/services/posts.service.js
@@ -32,13 +32,35 @@ class PostsService {
     return findPostsData;
   };
 
+  findOnePost = async (id) => {
+    const findOnePostData = await this.postsRespository.findOnePost(id);
+
+    if (findOnePostData.deleted_at !== 'null') {
+      throw new CustomError('삭제된 게시물 입니다.', 404);
+    }
+
+    return findOnePostData;
+  };
+
   updatePost = async (id, user_id, title, ingredient, recipe, food_img) => {
+    const findOnePostData = await this.postsRespository.findOnePost(id);
+
+    if (findOnePostData.deleted_at !== 'null') {
+      throw new CustomError('삭제된 게시물 입니다.', 404);
+    }
+
     const updatePostData = await this.postsRespository.updatePost(id, user_id, title, ingredient, recipe, food_img);
 
     return updatePostData;
   };
 
   deletePost = async (id, user_id) => {
+    const findOnePostData = await this.postsRespository.findOnePost(id);
+
+    if (findOnePostData.deleted_at !== 'null') {
+      throw new CustomError('이미 삭제된 게시물 입니다.', 404);
+    }
+
     const deletePostData = await this.postsRespository.deletePost(id, user_id);
 
     return deletePostData;
